requirements.txt
requests==2.31.0
pandas==2.1.4
click==8.1.7
python-dotenv==1.0.0
tabulate==0.9.0
